Deleting temp code so that i could merge
also commented out cam movement Player.idle();
