--- Assets/Scripts/Player/Attacking.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 49082a40486824f73bd3eea2c4bbfe92
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
