--- Assets/Scripts/Player/Move.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c038c35ce0959bd4e92c41714d751e82
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
