--- Assets/Scripts/Player/Player.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: bd4cc3ffe6a114d39a27ef77dacd3d32
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
