--- Assets/Scripts/Player/Reef/AttackAndMove.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3b29533a9a702b547aef909705224a57
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
