+++ Assets/Scripts/Utilities/CameraMovement.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class CameraMovement : MonoBehaviour
 {
-	public Player player;
+	public AttackAndMove player;
 	public bool Panning {get; private set;}
 	public enum Direction {Up, Down, Left, Right}
 
@@ -16,7 +16,7 @@ public class CameraMovement : MonoBehaviour
 	public IEnumerator MoveCamera(Direction direction)
 	{
 		Panning = true;
-		player.Idle();
+		//player.Idle();
 		int times = direction == Direction.Left || direction == Direction.Right ? 40 : 32;
 		Vector3 delta = new Vector3(0f,0f,0f);
 		if(direction == Direction.Left)
