+++ Assets/Sprites/Link-Reef/shield equip walks.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 2128788be5b81d446a648649b3400fca
+guid: c4acc502350f6034f8b7eb223ca2a94a
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
