+++ Assets/Sprites/Link-Reef/shield walks/shield walk left.psd.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f7bd09717d973504c8a5c1c190e27c85
+guid: ad5d99de5efabbd4e9ac93d4196b6437
 TextureImporter:
   internalIDToNameTable: []
   externalObjects: {}
