+++ Assets/Sprites/Link-Reef/shield walks/shield walk right.psd.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f647cacfb7ab9204084088d4ffd1d6b8
+guid: 4c6a4845d9e21964c990a2a78ac4d3ed
 TextureImporter:
   internalIDToNameTable: []
   externalObjects: {}
