+++ Assets/Sprites/Link-Reef/walks/idle down.anim.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: a1884f1c75026d14282429556a123684
+guid: 71f92738be0b5e446b57852b0b313276
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 0
