+++ Assets/Sprites/Link-Reef/walks/idle left.anim.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 8aae32ad550838e4da9927490b746f3f
+guid: 77012d8ab93a9bc4d9485cfd7d31709a
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 0
