+++ Assets/Sprites/Link-Reef/walks/idle right.anim.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 493f60d1ec9cc034ea7ceff77b2129a4
+guid: bb65a55516c3e644c8827fc3d8af2c25
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 0
