+++ Assets/Sprites/Link-Reef/walks/jump Left Anim.anim.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 3dfab31566175e84486137b2874b463f
+guid: 1f2345aea48509941a27a79fbba9e035
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 0
