+++ Assets/Sprites/Link-Reef/walks/jump Right Anim.anim.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: a9f247e72cc8f00458387ffd6f0a0382
+guid: 6970575fab404244089f810161555c29
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 0
