+++ Assets/Sprites/Link-Reef/walks/jump Up Anim.anim.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b4b9192aba755d4459a37b403964263a
+guid: cca45bd54399e8047ab9b44d2067798a
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 0
