+++ Assets/Sprites/Link-Reef/walks/slash down Anim.anim.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 1d7ab498127e2ec49a342303cfa707ee
+guid: dc0176754bf516e4fa856c973ed6d100
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 0
