+++ Assets/Sprites/Link-Reef/walks/slash left Anim.anim.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: af922bfd13c393a4a8c9e5df6a17cc5d
+guid: 60208372cdcc6d248850f875ad83d7da
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 0
