+++ Assets/Sprites/Link-Reef/walks/slash right Anim.anim.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 8a8993453da8e1e4dab0d1cd6d8fa45d
+guid: 5b4e914da7b936b47848b1ff3bc22acd
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 0
