+++ Assets/Sprites/Link-Reef/walks/slash up Anim.anim.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 8498769fb65826148a33a7f8f557d3dd
+guid: 5e4007e72de8e7b458d2713a3f1ea8b2
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 0
