+++ Assets/Sprites/Link-Reef/walks/walk up_0.controller.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: a0a2b0a56786814429c21eee01412e89
+guid: 4fae740ff05380947b9fde9205baa165
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 0
