+++ Assets/Sprites/link.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: d63fde317ac9baa4cbdfd1bb6e5bd6ce
-NativeFormatImporter:
+guid: 982a59c51f58a6e43a694e6d0580cc0a
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
-  mainObjectFileID: 0
   userData: 
   assetBundleName: 
   assetBundleVariant:
