+++ Assets/Sprites/link/jumps.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: d63fde317ac9baa4cbdfd1bb6e5bd6ce
-NativeFormatImporter:
+guid: dcaf34988cd68c0408882cee3691d9ba
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
-  mainObjectFileID: 0
   userData: 
   assetBundleName: 
   assetBundleVariant:
