+++ Assets/Sprites/link/player_controller.controller
@@ -311,6 +311,9 @@ AnimatorStateTransition:
   - m_ConditionMode: 2
     m_ConditionEvent: spinning
     m_EventTreshold: 0
+  - m_ConditionMode: 2
+    m_ConditionEvent: poking
+    m_EventTreshold: 0
   m_DstStateMachine: {fileID: 0}
   m_DstState: {fileID: 3527293615867702062}
   m_Solo: 0
