+++ Assets/Sprites/link/shield equip walks.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: d63fde317ac9baa4cbdfd1bb6e5bd6ce
-NativeFormatImporter:
+guid: 2128788be5b81d446a648649b3400fca
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
-  mainObjectFileID: 0
   userData: 
   assetBundleName: 
   assetBundleVariant:
