+++ Assets/Sprites/link/slash.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: d63fde317ac9baa4cbdfd1bb6e5bd6ce
-NativeFormatImporter:
+guid: f37ffc16b255fe74683d2f7efc9bb5a1
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
-  mainObjectFileID: 0
   userData: 
   assetBundleName: 
   assetBundleVariant:
