+++ Assets/Sprites/link/slash/spin slash/spin right slash.anim
@@ -22,32 +22,32 @@ AnimationClip:
     - time: 0
       value: {fileID: 1836898549899966914, guid: ea82ca487d3a9d846a51aaee220d63c6,
         type: 3}
-    - time: 0.083333336
+    - time: 0.11111111
       value: {fileID: -3703305374327498108, guid: ea82ca487d3a9d846a51aaee220d63c6,
         type: 3}
-    - time: 0.16666667
+    - time: 0.22222222
       value: {fileID: -2312491225392559394, guid: ea82ca487d3a9d846a51aaee220d63c6,
         type: 3}
-    - time: 0.25
+    - time: 0.33333334
       value: {fileID: 6830862982645221754, guid: ea82ca487d3a9d846a51aaee220d63c6,
         type: 3}
-    - time: 0.33333334
+    - time: 0.44444445
       value: {fileID: 8721817098185123453, guid: ea82ca487d3a9d846a51aaee220d63c6,
         type: 3}
-    - time: 0.41666666
+    - time: 0.5555556
       value: {fileID: 1794222831342123273, guid: ea82ca487d3a9d846a51aaee220d63c6,
         type: 3}
-    - time: 0.5
+    - time: 0.6666667
       value: {fileID: -7114454125260075053, guid: ea82ca487d3a9d846a51aaee220d63c6,
         type: 3}
-    - time: 0.5833333
+    - time: 0.7777778
       value: {fileID: -4201152883117719921, guid: ea82ca487d3a9d846a51aaee220d63c6,
         type: 3}
     attribute: m_Sprite
     path: 
     classID: 212
     script: {fileID: 0}
-  m_SampleRate: 12
+  m_SampleRate: 9
   m_WrapMode: 0
   m_Bounds:
     m_Center: {x: 0, y: 0, z: 0}
@@ -75,7 +75,7 @@ AnimationClip:
     m_AdditiveReferencePoseClip: {fileID: 0}
     m_AdditiveReferencePoseTime: 0
     m_StartTime: 0
-    m_StopTime: 0.6666666
+    m_StopTime: 0.8888889
     m_OrientationOffsetY: 0
     m_Level: 0
     m_CycleOffset: 0
