+++ Assets/Sprites/link/slash/spin slash/spin slash left.anim
@@ -22,32 +22,32 @@ AnimationClip:
     - time: 0
       value: {fileID: -2218482595022607518, guid: 6fa2378beba6f7a4aacb201af03e5fe6,
         type: 3}
-    - time: 0.083333336
+    - time: 0.11111111
       value: {fileID: 4137003120548311776, guid: 6fa2378beba6f7a4aacb201af03e5fe6,
         type: 3}
-    - time: 0.16666667
+    - time: 0.22222222
       value: {fileID: 8149100650908649313, guid: 6fa2378beba6f7a4aacb201af03e5fe6,
         type: 3}
-    - time: 0.25
+    - time: 0.33333334
       value: {fileID: -8009223636436758353, guid: 6fa2378beba6f7a4aacb201af03e5fe6,
         type: 3}
-    - time: 0.33333334
+    - time: 0.44444445
       value: {fileID: -6756125980404665516, guid: 6fa2378beba6f7a4aacb201af03e5fe6,
         type: 3}
-    - time: 0.41666666
+    - time: 0.5555556
       value: {fileID: -2020986857461159255, guid: 6fa2378beba6f7a4aacb201af03e5fe6,
         type: 3}
-    - time: 0.5
+    - time: 0.6666667
       value: {fileID: 139074016337346293, guid: 6fa2378beba6f7a4aacb201af03e5fe6,
         type: 3}
-    - time: 0.5833333
+    - time: 0.7777778
       value: {fileID: -7785348479374994533, guid: 6fa2378beba6f7a4aacb201af03e5fe6,
         type: 3}
     attribute: m_Sprite
     path: 
     classID: 212
     script: {fileID: 0}
-  m_SampleRate: 12
+  m_SampleRate: 9
   m_WrapMode: 0
   m_Bounds:
     m_Center: {x: 0, y: 0, z: 0}
@@ -75,7 +75,7 @@ AnimationClip:
     m_AdditiveReferencePoseClip: {fileID: 0}
     m_AdditiveReferencePoseTime: 0
     m_StartTime: 0
-    m_StopTime: 0.6666666
+    m_StopTime: 0.8888889
     m_OrientationOffsetY: 0
     m_Level: 0
     m_CycleOffset: 0
