+++ Assets/Sprites/link/slash/spin slash/spin slash up.anim
@@ -22,32 +22,32 @@ AnimationClip:
     - time: 0
       value: {fileID: 7182612611632546498, guid: 8acd2d49e4f5fb94a86fdffa59809226,
         type: 3}
-    - time: 0.083333336
+    - time: 0.11111111
       value: {fileID: -1730691248723393097, guid: 8acd2d49e4f5fb94a86fdffa59809226,
         type: 3}
-    - time: 0.16666667
+    - time: 0.22222222
       value: {fileID: 272467162767372954, guid: 8acd2d49e4f5fb94a86fdffa59809226,
         type: 3}
-    - time: 0.25
+    - time: 0.33333334
       value: {fileID: 8637601313527524606, guid: 8acd2d49e4f5fb94a86fdffa59809226,
         type: 3}
-    - time: 0.33333334
+    - time: 0.44444445
       value: {fileID: -2295672382817120827, guid: 8acd2d49e4f5fb94a86fdffa59809226,
         type: 3}
-    - time: 0.41666666
+    - time: 0.5555556
       value: {fileID: 5065971109391818294, guid: 8acd2d49e4f5fb94a86fdffa59809226,
         type: 3}
-    - time: 0.5
+    - time: 0.6666667
       value: {fileID: 5355046946204556173, guid: 8acd2d49e4f5fb94a86fdffa59809226,
         type: 3}
-    - time: 0.5833333
+    - time: 0.7777778
       value: {fileID: 2108290949444931376, guid: 8acd2d49e4f5fb94a86fdffa59809226,
         type: 3}
     attribute: m_Sprite
     path: 
     classID: 212
     script: {fileID: 0}
-  m_SampleRate: 12
+  m_SampleRate: 9
   m_WrapMode: 0
   m_Bounds:
     m_Center: {x: 0, y: 0, z: 0}
@@ -75,7 +75,7 @@ AnimationClip:
     m_AdditiveReferencePoseClip: {fileID: 0}
     m_AdditiveReferencePoseTime: 0
     m_StartTime: 0
-    m_StopTime: 0.6666666
+    m_StopTime: 0.8888889
     m_OrientationOffsetY: 0
     m_Level: 0
     m_CycleOffset: 0
